requirements.txt
Django==5.2.4
djangorestframework==3.16.1
reportlab==4.4.10
gunicorn
whitenoise